moved products services to /products folder
