+++ src/frontend/services/products/createNewProduct.ts
@@ -1,9 +1,12 @@
-import publicAxiosInstance from '../api/axios-api'
+import publicAxiosInstance from '../../api/axios-api'
 type ProductData = {
   name: string
   description: string
   category: string
   price: number
+  storeId: string
+  quantity: number
+  minQuantity: number
 
 }
 export const createNewProduct = async (product: ProductData) => {
