+++ src/frontend/services/products/deleteProducts.ts
@@ -1,4 +1,4 @@
-import publicAxiosInstance from '../api/axios-api'
+import publicAxiosInstance from '../../api/axios-api'
 export const deleteProduct = async (productId: string) => {
   const response = await publicAxiosInstance.delete(
     `/products/${productId}`,
