+++ src/frontend/services/products/filterProducts.ts
@@ -1,4 +1,4 @@
-import { Product } from '../../shared/schemas/product.type'
+import { Product } from '../../../shared/schemas/product.type'
 
 export const filterProducts = (products: Product[], search: string) => {
   return products?.filter((product) => {
