+++ src/frontend/services/products/getProducts.ts
@@ -1,4 +1,4 @@
-import publicAxiosInstance from '../api/axios-api'
+import publicAxiosInstance from '../../api/axios-api'
 
 export const getProducts = async () => {
   const response = await publicAxiosInstance.get('/products')
