+++ src/frontend/services/products/updateProduct.ts
@@ -1,5 +1,5 @@
-import { Product } from '../../shared/schemas/product.type'
-import publicAxiosInstance from '../api/axios-api'
+import { Product } from '../../../shared/schemas/product.type'
+import publicAxiosInstance from '../../api/axios-api'
 
 export const updateProduct = async ({ id, name, price, description, category }: Product) => {
   const response = await publicAxiosInstance.put(`/products/${id}`, {
